external user image issue

+++ api/class/users.class.php
@@ -452,11 +452,13 @@ class users {
                     $this->_db->where('iUserId',$id);
                     $data = $this->_db->getOne('tms_users');
                     $image = $data['vProfilePic'];
-                    if(isset($image)){
+                    if(isset($image) && $image != ''){
                         $path = "../../uploads/profilePic/";
                         $images = glob($path.$image);
                         if($images){
-                            unlink($path.$image);
+                            if(file_exists($path.$image)){
+                                unlink($path.$image);
+                            }
                         }else{
                             $image = $data['vProfilePic'];
                         }
